backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
httpx
ruff
black
mypy
types-passlib
types-python-jose
pydantic-settings
slowapi
pytest
pytest-asyncio
httpx
alembic
psycopg2-binary

